G92 T axis value missing offset in 4-axis mode in GCV
When clicking on the zero-axes button (to the right of the Home
button), a G92 command of the form

  G92 x0 y0 z0 a

would be incorrectly emitted.  This change moves the toLower() method
so that it is called on the QString instance of the axis name
attribute directly rather than on the result of the append() method.